Move network layer from service to system tier
Relocate all networking implementation (SecureService, overlay Manager,
Tor/I2P/Tailscale providers) from service/net/ to system/net/ and unify
the two boot components into a single system-level component. Networking
is foundational infrastructure, not an application service.
