+++ boot/components/service/all.go
@@ -16,7 +16,6 @@ import (
 
 func All() []boot.Component {
 	components := []boot.Component{
-		Network(),
 		fs.Directory(),
 		fs.Embed(),
 		Template(),
